+++ openmdao/core/system.py
@@ -5,14 +5,14 @@ import hashlib
 import pathlib
 import time
 import functools
+import textwrap
 
 from contextlib import contextmanager
-from collections import defaultdict
+from collections import defaultdict, namedtuple
 from itertools import chain
 from enum import IntEnum
-
+from io import StringIO
 from fnmatch import fnmatchcase
-
 from numbers import Integral
 
 import numpy as np
@@ -28,7 +28,7 @@ from openmdao.utils.options_dictionary import OptionsDictionary
 from openmdao.utils.record_util import create_local_meta, check_path, has_match
 from openmdao.utils.units import is_compatible, unit_conversion, simplify_unit
 from openmdao.utils.variable_table import write_var_table, NA
-from openmdao.utils.array_utils import evenly_distrib_idxs, shape_to_len
+from openmdao.utils.array_utils import evenly_distrib_idxs, shape_to_len, safe_norm
 from openmdao.utils.name_maps import name2abs_name, name2abs_names
 from openmdao.utils.coloring import _compute_coloring, Coloring, \
     _STD_COLORING_FNAME, _DEF_COMP_SPARSITY_ARGS, _ColSparsityJac
@@ -39,10 +39,11 @@ from openmdao.utils.om_warnings import issue_warning, \
 from openmdao.utils.general_utils import determine_adder_scaler, \
     format_as_float_or_array, all_ancestors, match_prom_or_abs, \
     ensure_compatible, env_truthy, make_traceback, _is_slicer_op, _wrap_comm, _unwrap_comm, \
-    _om_dump, SystemMetaclass
+    _om_dump, SystemMetaclass, add_border
 from openmdao.utils.file_utils import _get_outputs_dir
 from openmdao.approximation_schemes.complex_step import ComplexStep
 from openmdao.approximation_schemes.finite_difference import FiniteDifference
+from openmdao.visualization.tables.table_builder import generate_table
 
 _empty_frozen_set = frozenset()
 
@@ -6790,3 +6791,724 @@ class System(object, metaclass=SystemMetaclass):
             if path not in seen:
                 seen.add(path)
                 yield path
+
+    def _deriv_display(self, err_iter, derivatives, rel_error_tol, abs_error_tol, out_stream,
+                       fd_opts, totals=False, show_only_incorrect=False, lcons=None, sort=False):
+        """
+        Compute the relative and absolute errors in the given derivatives and print to the out_stream.
+
+        Parameters
+        ----------
+        err_iter : iterator
+            Iterator that yields tuples of the form (key, fd_norm, fd_opts, directional, above_abs,
+            above_rel, inconsistent) for each subjac.
+        derivatives : dict
+            Dictionary containing derivative information keyed by (of, wrt).
+        rel_error_tol : float
+            Relative error tolerance.
+        abs_error_tol : float
+            Absolute error tolerance.
+        out_stream : file-like object
+                Where to send human readable output.
+                Set to None to suppress.
+        totals : bool or _TotalJacInfo
+            Set to _TotalJacInfo if we are doing check_totals to skip a bunch of stuff.
+        show_only_incorrect : bool, optional
+            Set to True if output should print only the subjacs found to be incorrect.
+        lcons : list or None
+            For total derivatives only, list of outputs that are actually linear constraints.
+        sort : bool
+            If True, sort subjacobian keys alphabetically.
+        """
+        from openmdao.core.component import Component
+
+        # Match header to appropriate type.
+        if isinstance(self, Component):
+            sys_type = 'Component'
+        else:
+            sys_type = 'Group'
+
+        sys_name = self.pathname
+        sys_class_name = type(self).__name__
+
+        if totals:
+            sys_name = 'Full Model'
+
+        num_bad_jacs = 0  # Keep track of number of bad derivative values for each component
+
+        # Need to capture the output of a component's derivative
+        # info so that it can be used if that component is the
+        # worst subjac. That info is printed at the bottom of all the output
+        sys_buffer = StringIO()
+
+        if out_stream is not None:
+            if totals:
+                title = "Total Derivatives"
+            else:
+                title = f"{sys_type}: {sys_class_name} '{sys_name}'"
+
+            print(f"{add_border(title, '-')}\n", file=sys_buffer)
+
+        for key, fd_norm, fd_opts, directional, above_abs, above_rel, inconsistent in err_iter:
+
+            if above_abs or above_rel or inconsistent:
+                num_bad_jacs += 1
+
+            if out_stream is None:
+                continue
+
+            of, wrt = key
+            derivative_info = derivatives[key]
+
+            # Informative output for responses that were declared with an index.
+            indices = derivative_info.get('indices')
+            if indices is not None:
+                of = f'{of} (index size: {indices})'
+
+            # need this check because if directional may be list
+            if isinstance(wrt, str):
+                wrt = f"'{wrt}'"
+            if isinstance(of, str):
+                of = f"'{of}'"
+
+            if directional:
+                wrt = f"(d){wrt}"
+
+            abs_errs = derivative_info['abs error']
+            rel_errs = derivative_info['rel error']
+            magnitudes = derivative_info['magnitude']
+            steps = derivative_info['steps']
+
+            Jfor = derivative_info.get('J_fwd')
+            if Jfor is None:
+                Jrev = derivative_info['J_rev']
+
+            if len(steps) > 1:
+                stepstrs = [f", step={step}" for step in steps]
+            else:
+                stepstrs = [""]
+
+            if fd_norm == 0.:
+                if magnitudes[0].forward is None:
+                    divname = 'Jrev'
+                else:
+                    divname = 'Jfor'
+            else:
+                divname = 'Jfd'
+
+            if out_stream:
+                # Magnitudes
+                sys_buffer.write(f"  {sys_name}: {of} wrt {wrt}")
+                if not isinstance(of, tuple) and lcons and of.strip("'") in lcons:
+                    sys_buffer.write(" (Linear constraint)")
+
+                sys_buffer.write('\n')
+                if magnitudes[0].forward is not None:
+                    sys_buffer.write(f'     Forward Magnitude: {magnitudes[0].forward:.6e}\n')
+
+                if magnitudes[0].reverse is not None:
+                    sys_buffer.write(f'     Reverse Magnitude: {magnitudes[0].reverse:.6e}\n')
+
+                fd_desc = f"{fd_opts['method']}:{fd_opts['form']}"
+                for i in range(len(magnitudes)):
+                    sys_buffer.write(f'          Fd Magnitude: '
+                                        f'{magnitudes[i].fd:.6e} ({fd_desc}{stepstrs[i]})\n')
+                sys_buffer.write('\n')
+
+                for i in range(len(magnitudes)):
+                    # Absolute Errors
+                    if directional:
+                        if totals and abs_errs[i].forward is not None:
+                            err = _format_error(abs_errs[i].forward, abs_error_tol)
+                            sys_buffer.write(f'    Absolute Error (Jfor - Jfd){stepstrs[i]} : '
+                                                f'{err}\n')
+
+                        if abs_errs[i].reverse is not None:
+                            err = _format_error(abs_errs[i].reverse, abs_error_tol)
+                            sys_buffer.write(f'    Absolute Error ([rev, fd] Dot Product Test)'
+                                                f'{stepstrs[i]} : {err}\n')
+                    else:
+                        if abs_errs[i].forward is not None:
+                            err = _format_error(abs_errs[i].forward, abs_error_tol)
+                            sys_buffer.write(f'    Absolute Error (Jfor - Jfd){stepstrs[i]} : '
+                                                f'{err}\n')
+
+                        if abs_errs[i].reverse is not None:
+                            err = _format_error(abs_errs[i].reverse, abs_error_tol)
+                            sys_buffer.write(f'    Absolute Error (Jrev - Jfd){stepstrs[i]} : '
+                                                f'{err}\n')
+
+                if directional:
+                    if abs_errs[0].forward_reverse is not None:
+                        err = _format_error(abs_errs[0].forward_reverse, abs_error_tol)
+                        sys_buffer.write('    Absolute Error ([rev, for] Dot Product Test) : '
+                                            f'{err}\n')
+                else:
+                    if abs_errs[0].forward_reverse is not None:
+                        err = _format_error(abs_errs[0].forward_reverse, abs_error_tol)
+                        sys_buffer.write(f'    Absolute Error (Jrev - Jfor) : {err}\n')
+
+            sys_buffer.write('\n')
+
+            for i in range(len(magnitudes)):
+                # Relative Errors
+                if out_stream:
+                    if directional:
+                        if totals and rel_errs[i].forward is not None:
+                            err = _format_error(rel_errs[i].forward, rel_error_tol)
+                            sys_buffer.write(f'    Relative Error (Jfor - Jfd) / {divname}'
+                                                f'{stepstrs[i]} : {err}\n')
+
+                        if rel_errs[i].reverse is not None:
+                            err = _format_error(rel_errs[i].reverse, rel_error_tol)
+                            sys_buffer.write(f'    Relative Error ([rev, fd] Dot Product Test) '
+                                                f'/ {divname}{stepstrs[i]} : {err}\n')
+                    else:
+                        if rel_errs[i].forward is not None:
+                            err = _format_error(rel_errs[i].forward, rel_error_tol)
+                            sys_buffer.write(f'    Relative Error (Jfor - Jfd) / {divname}'
+                                                f'{stepstrs[i]} : {err}\n')
+
+                        if rel_errs[i].reverse is not None:
+                            err = _format_error(rel_errs[i].reverse, rel_error_tol)
+                            sys_buffer.write(f'    Relative Error (Jrev - Jfd) / {divname}'
+                                                f'{stepstrs[i]} : {err}\n')
+
+            if out_stream:
+                if directional:
+                    if rel_errs[0].forward_reverse is not None:
+                        err = _format_error(rel_errs[0].forward_reverse, rel_error_tol)
+                        sys_buffer.write(f'    Relative Error ([rev, for] Dot Product Test) / '
+                                            f'{divname} : {err}\n')
+                else:
+                    if rel_errs[0].forward_reverse is not None:
+                        err = _format_error(rel_errs[0].forward_reverse, rel_error_tol)
+                        sys_buffer.write(f'    Relative Error (Jrev - Jfor) / {divname} : '
+                                            f'{err}\n')
+
+            if inconsistent:
+                sys_buffer.write('\n    * Inconsistent value across ranks *\n')
+
+            comm = self._problem_meta['comm']
+            if MPI and comm.size > 1:
+                sys_buffer.write(f'\n    MPI Rank {comm.rank}\n')
+            sys_buffer.write('\n')
+
+            with np.printoptions(linewidth=240):
+                # Raw Derivatives
+                if magnitudes[0].forward is not None:
+                    if directional:
+                        sys_buffer.write('    Directional Derivative (Jfor)')
+                    else:
+                        sys_buffer.write('    Raw Forward Derivative (Jfor)')
+                    Jstr = textwrap.indent(str(Jfor), '    ')
+                    sys_buffer.write(f"\n{Jstr}\n\n")
+
+                fdtype = fd_opts['method'].upper()
+
+                if magnitudes[0].reverse is not None:
+                    if directional:
+                        if totals:
+                            sys_buffer.write('    Directional Derivative (Jrev) Dot Product')
+                        else:
+                            sys_buffer.write('    Directional Derivative (Jrev)')
+                    else:
+                        sys_buffer.write('    Raw Reverse Derivative (Jrev)')
+                    Jstr = textwrap.indent(str(Jrev), '    ')
+                    sys_buffer.write(f"\n{Jstr}\n\n")
+
+                try:
+                    fds = derivative_info['J_fd']
+                except KeyError:
+                    fds = [0.]
+
+                for i in range(len(magnitudes)):
+                    fd = fds[i]
+
+                    if directional:
+                        if totals and magnitudes[i].reverse is not None:
+                            sys_buffer.write(f'    Directional {fdtype} Derivative (Jfd) '
+                                                f'Dot Product{stepstrs[i]}\n    {fd}\n\n')
+                        else:
+                            sys_buffer.write(f"    Directional {fdtype} Derivative (Jfd)"
+                                                f"{stepstrs[i]}\n    {fd}\n\n")
+                    else:
+                        Jstr = textwrap.indent(str(fd), '    ')
+                        sys_buffer.write(f"    Raw {fdtype} Derivative (Jfd){stepstrs[i]}"
+                                            f"\n{Jstr}\n\n")
+
+            sys_buffer.write(' -' * 30 + '\n')
+
+        # End of for of, wrt in sorted_keys
+
+        if out_stream is not None and (not show_only_incorrect or num_bad_jacs > 0):
+            out_stream.write(sys_buffer.getvalue())
+
+    def _deriv_display_compact(self, err_iter, derivatives, rel_error_tol, abs_error_tol,
+                               out_stream, fd_opts, totals=False,
+                               show_only_incorrect=False, lcons=None, sort=False):
+        """
+        Compute relative and absolute errors in the given derivatives and print to out_stream.
+
+        Display is in a compact tabular format.
+
+        Parameters
+        ----------
+        err_iter : iterator
+            Iterator that yields tuples of the form (key, fd_norm, fd_opts, directional, above_abs,
+            above_rel, inconsistent) for each subjac.
+        derivatives : dict
+            Dictionary containing derivative information keyed by (of, wrt).
+        rel_error_tol : float
+            Relative error tolerance.
+        abs_error_tol : float
+            Absolute error tolerance.
+        out_stream : file-like object
+                Where to send human readable output.
+                Set to None to suppress.
+        fd_opts : dict
+            Dictionary containing the options for the approximation.
+        totals : bool or _TotalJacInfo
+            Set to _TotalJacInfo if we are doing check_totals to skip a bunch of stuff.
+        show_only_incorrect : bool, optional
+            Set to True if output should print only the subjacs found to be incorrect.
+        lcons : list or None
+            For total derivatives only, list of outputs that are actually linear constraints.
+        sort : bool
+            If True, sort subjacobian keys alphabetically.
+        """
+        from openmdao.core.component import Component
+
+        # Match header to appropriate type.
+        if isinstance(self, Component):
+            sys_type = 'Component'
+        else:
+            sys_type = 'Group'
+
+        sys_name = self.pathname
+        sys_class_name = type(self).__name__
+        matrix_free = self.matrix_free
+
+        if totals:
+            sys_name = 'Full Model'
+
+        num_bad_jacs = 0  # Keep track of number of bad derivative values for each component
+
+        # Need to capture the output of a component's derivative
+        # info so that it can be used if that component is the
+        # worst subjac. That info is printed at the bottom of all the output
+        sys_buffer = StringIO()
+
+        if out_stream is not None:
+            if totals:
+                title = "Total Derivatives"
+            else:
+                title = f"{sys_type}: {sys_class_name} '{sys_name}'"
+
+            print(f"{add_border(title, '-')}\n", file=sys_buffer)
+
+            table_data = []
+            worst_subjac = None
+
+        for key, fd_norm, fd_opts, directional, above_abs, above_rel, inconsistent in err_iter:
+
+            if above_abs or above_rel or inconsistent:
+                num_bad_jacs += 1
+
+            if out_stream is None:
+                continue
+
+            of, wrt = key
+            derivative_info = derivatives[key]
+
+            # Informative output for responses that were declared with an index.
+            indices = derivative_info.get('indices')
+            if indices is not None:
+                of = f'{of} (index size: {indices})'
+
+            # need this check because if directional may be list
+            if isinstance(wrt, str):
+                wrt = f"'{wrt}'"
+            if isinstance(of, str):
+                of = f"'{of}'"
+
+            if directional:
+                wrt = f"(d){wrt}"
+
+            err_desc = []
+            if above_abs:
+                err_desc.append('>ABS_TOL')
+            if above_rel:
+                err_desc.append('>REL_TOL')
+            if inconsistent:
+                err_desc.append('<RANK INCONSISTENT>')
+            err_desc = ' '.join(err_desc)
+
+            abs_errs = derivative_info['abs error']
+            rel_errs = derivative_info['rel error']
+            magnitudes = derivative_info['magnitude']
+            steps = derivative_info['steps']
+
+            for magnitude, abs_err, rel_err, step in zip(magnitudes, abs_errs, rel_errs, steps):
+                if magnitude.reverse is not None:
+                    calc_mag = magnitude.reverse
+                    calc_abs = abs_err.reverse
+                    calc_rel = rel_err.reverse
+
+                # use forward even if both fwd and rev are defined
+                if magnitude.forward is not None:
+                    calc_mag = magnitude.forward
+                    calc_abs = abs_err.forward
+                    calc_rel = rel_err.forward
+
+                if totals:
+                    if len(steps) > 1:
+                        table_data.append([of, wrt, step, calc_mag, magnitude.fd,
+                                            calc_abs, calc_rel, err_desc])
+                    else:
+                        table_data.append([of, wrt, calc_mag, magnitude.fd,
+                                            calc_abs, calc_rel, err_desc])
+                else:
+                    if matrix_free:
+                        if len(steps) > 1:
+                            table_data.append([of, wrt, step, magnitude.forward,
+                                                magnitude.reverse, magnitude.fd,
+                                                abs_err.forward, abs_err.reverse,
+                                                abs_err.forward_reverse, rel_err.forward,
+                                                rel_err.reverse, rel_err.forward_reverse,
+                                                err_desc])
+                        else:
+                            table_data.append([of, wrt, magnitude.forward,
+                                                magnitude.reverse, magnitude.fd,
+                                                abs_err.forward, abs_err.reverse,
+                                                abs_err.forward_reverse, rel_err.forward,
+                                                rel_err.reverse, rel_err.forward_reverse,
+                                                err_desc])
+                    else:
+                        if len(steps) > 1:
+                            table_data.append([of, wrt, step, magnitude.forward,
+                                                magnitude.fd, abs_err.forward,
+                                                rel_err.forward, err_desc])
+                        else:
+                            table_data.append([of, wrt, magnitude.forward, magnitude.fd,
+                                                abs_err.forward, rel_err.forward,
+                                                err_desc])
+                        assert abs_err.forward_reverse is None
+                        assert rel_err.forward_reverse is None
+                        assert abs_err.reverse is None
+                        assert rel_err.reverse is None
+
+                    # See if this component has the greater error in the derivative computation
+                    # compared to the other components so far
+                    for err in rel_err[:2]:
+                        if err is None or np.isnan(err):
+                            continue
+
+                        if worst_subjac is None or err > worst_subjac[0]:
+                            worst_subjac = (err, table_data[-1])
+
+        headers = []
+        if out_stream is not None:
+            if table_data:
+                headers = ["of '<variable>'", "wrt '<variable>'"]
+                if len(steps) > 1:
+                    headers.append('step')
+
+                if matrix_free:
+                    headers.extend(['fwd mag.', 'rev mag.', 'check mag.', 'a(fwd-chk)',
+                                    'a(rev-chk)', 'a(fwd-rev)', 'r(fwd-chk)', 'r(rev-chk)',
+                                    'r(fwd-rev)', 'error desc'])
+                else:
+                    headers.extend(['calc mag.', 'check mag.', 'a(cal-chk)', 'r(cal-chk)',
+                                    'error desc'])
+
+                _print_deriv_table(table_data, headers, sys_buffer)
+
+            if not show_only_incorrect or num_bad_jacs > 0:
+                out_stream.write(sys_buffer.getvalue())
+
+        if worst_subjac is None:
+            return None
+
+        return worst_subjac + (headers,) if out_stream is not None else None
+
+
+_ErrorTuple = namedtuple('ErrorTuple', ['forward', 'reverse', 'forward_reverse'])
+_MagnitudeTuple = namedtuple('MagnitudeTuple', ['forward', 'reverse', 'fd'])
+
+
+def _print_deriv_table(table_data, headers, out_stream):
+    """
+    Print a table of derivatives.
+
+    Parameters
+    ----------
+    table_data : list
+        List of lists containing the table data.
+    headers : list
+        List of column headers.
+    out_stream : file-like object
+        Where to send human readable output.
+    """
+    num_col_meta = {'format': '{: 1.4e}'}
+    column_meta = [{}, {}]
+    column_meta.extend([num_col_meta.copy() for _ in range(len(headers) - 3)])
+    column_meta.append({})
+    print(generate_table(table_data, headers=headers, tablefmt='grid',
+                         column_meta=column_meta, missing_val='n/a'), file=out_stream)
+
+
+def _compute_deriv_errors(derivative_info, matrix_free, directional, totals):
+    """
+    Compute the errors between derivatives that were computed using different modes or methods.
+
+    Error information in the derivative_info dict is updated by this function.
+
+    Parameters
+    ----------
+    derivative_info : dict
+        Metadata dict corresponding to a particular (of, wrt) pair.
+    matrix_free : bool
+        True if the current dirivatives are computed in a matrix free manner.
+    directional : bool
+        True if the current dirivtives are directional.
+    totals : bool or _TotalJacInfo
+        _TotalJacInfo if the current derivatives are total derivatives.
+
+    Returns
+    -------
+    float
+        The norm of the FD jacobian.
+    """
+    nan = float('nan')
+
+    Jforward = derivative_info.get('J_fwd')
+    Jreverse = derivative_info.get('J_rev')
+    forward = Jforward is not None
+    reverse = Jreverse is not None
+
+    rev_norm = fwd_norm = fwd_rev_error = None
+    calc_norm = 0.
+    if reverse:
+        rev_norm = calc_norm = safe_norm(Jreverse)
+    if forward:
+        fwd_norm = calc_norm = safe_norm(Jforward)
+
+    try:
+        fdinfo = derivative_info['J_fd']
+        steps = derivative_info['steps']
+    except KeyError:
+        # this can happen when a partial is not declared, which means it should be zero
+        fdinfo = (np.zeros(1),)
+        steps = (None,)
+
+    if matrix_free:
+        derivative_info['matrix_free'] = True
+        if directional:
+            if forward and reverse:
+                mhatdotm, dhatdotd = derivative_info['directional_fwd_rev']
+                fwd_rev_error = safe_norm(mhatdotm - dhatdotd)
+            else:
+                fwd_rev_error = None
+        elif not totals:
+            fwd_rev_error = safe_norm(Jforward - Jreverse)
+
+    derivative_info['abs error'] = []
+    derivative_info['rel error'] = []
+    derivative_info['magnitude'] = []
+    derivative_info['steps'] = []
+    fdnorms = []
+
+    for i, fd in enumerate(fdinfo):
+        step = steps[i]
+        fd_norm = safe_norm(fd)
+        fdnorms.append(fd_norm)
+
+        fwd_error = rev_error = None
+
+        if reverse and not directional:
+            rev_error = safe_norm(Jreverse - fd)
+
+        if forward:
+            fwd_error = safe_norm(Jforward - fd)
+
+        if directional:
+            if reverse:
+                mhatdotm, dhatdotd = derivative_info['directional_fd_rev'][i]
+                rev_error = safe_norm(mhatdotm - dhatdotd)
+                if not totals:
+                    rev_norm = None
+            if forward and totals:
+                mhatdotm, dhatdotd = derivative_info['directional_fd_fwd'][i]
+                fwd_error = safe_norm(mhatdotm - dhatdotd)
+
+        derivative_info['abs error'].append(_ErrorTuple(fwd_error, rev_error, fwd_rev_error))
+        derivative_info['magnitude'].append(_MagnitudeTuple(fwd_norm, rev_norm, fd_norm))
+        derivative_info['steps'].append(step)
+
+        # If fd_norm is zero, let's use calc_norm as the divisor for the relative
+        # error check. That way we don't accidentally squelch a legitimate problem.
+        div_norm = fd_norm if fd_norm != 0. else calc_norm
+
+        if div_norm == 0.:
+            derivative_info['rel error'].append(_ErrorTuple(None if fwd_error is None else nan,
+                                                            None if rev_error is None else nan,
+                                                            None if fwd_rev_error is None else nan))
+        else:
+            if matrix_free and not totals:
+                derivative_info['rel error'].append(_ErrorTuple(fwd_error / div_norm,
+                                                                rev_error / div_norm,
+                                                                fwd_rev_error / div_norm))
+            else:
+                derivative_info['rel error'].append(_ErrorTuple(
+                    None if fwd_error is None else fwd_error / div_norm,
+                    None if rev_error is None else rev_error / div_norm,
+                    None if fwd_rev_error is None else fwd_rev_error / div_norm))
+
+    return np.max(fdnorms)
+
+
+def _errors_above_tol(deriv_info, abs_error_tol, rel_error_tol):
+    """
+    Return if either abs or rel tolerances are violated when comparing a group of derivatives.
+
+    Parameters
+    ----------
+    deriv_info : dict
+        Metadata dict corresponding to a particular (of, wrt) pair.
+    abs_error_tol : float
+        Absolute error tolerance.
+    rel_error_tol : float
+        Relative error tolerance.
+
+    Returns
+    -------
+    bool
+        True if absolute tolerance is violated.
+    bool
+        True if relative tolerance is violated.
+    """
+    abs_errs = deriv_info['abs error']
+    rel_errs = deriv_info['rel error']
+
+    above_abs = above_rel = False
+
+    for abs_err in abs_errs:
+        for error in abs_err:
+            if error is not None and not np.isnan(error) and error >= abs_error_tol:
+                above_abs = True
+                break
+        if above_abs:
+            break
+
+    for rel_err in rel_errs:
+        for error in rel_err:
+            if error is not None and not np.isnan(error) and error >= rel_error_tol:
+                above_rel = True
+                break
+        if above_rel:
+            break
+
+    return above_abs, above_rel
+
+
+def _iter_derivs(derivatives, show_only_incorrect, all_fd_opts, totals, nondep_derivs,
+                 matrix_free, abs_error_tol=1e-6, rel_error_tol=1e-6, incon_keys=()):
+    """
+    Iterate over all of the derivatives.
+
+    If show_only_incorrect is True, only the derivatives with abs or rel errors outside of
+    tolerance or derivatives wrt serial variables that are inconsistent across ranks will be
+    returned.
+
+    Parameters
+    ----------
+    derivatives : dict
+        Dict of metadata for derivative groups, keyed on (of, wrt) pairs.
+    show_only_incorrect : bool
+        If True, yield only derivatives with errors outside of tolerance.
+    all_fd_opts : dict
+        Dictionary containing the options for the approximation.
+    totals : bool or _TotalJacInfo
+        Set to _TotalJacInfo if we are doing check_totals to skip a bunch of stuff.
+    nondep_derivs : set
+        Contains the of/wrt keys that are declared not dependent or not declared at all.
+    matrix_free : bool
+        True if the system computes matrix free derivatives.
+    abs_error_tol : float
+        Absolute error tolerance.
+    rel_error_tol : float
+        Relative error tolerance.
+    incon_keys : set or tuple
+        Keys where there are serial d_inputs variables that are inconsistent across processes.
+
+    Yields
+    ------
+    tuple
+        The (of, wrt) pair for the current derivatives being compared.
+    float
+        The FD norm.
+    dict
+        The FD options.
+    bool
+        True if the current derivatives are directional.
+    bool
+        True if the differences for the current derivatives are above the absolute error tolerance.
+    bool
+        True if the differences for the current derivatives are above the relative error tolerance.
+    bool
+        True if the current derivative was computed where some serial d_inputs variables were not
+        consistent across processes.
+    """
+    # Sorted keys ensures deterministic ordering
+    sorted_keys = sorted(derivatives)
+
+    for key in sorted_keys:
+        _, wrt = key
+
+        inconsistent = False
+        derivative_info = derivatives[key]
+
+        if totals:
+            fd_opts = all_fd_opts
+        else:
+            fd_opts = all_fd_opts[wrt]
+            if key in incon_keys:
+                inconsistent = True
+
+        directional = bool(fd_opts) and fd_opts.get('directional')
+
+        fd_norm = _compute_deriv_errors(derivative_info, matrix_free, directional, totals)
+
+        # Skip printing the non-dependent keys if the derivatives are fine.
+        if key in nondep_derivs and fd_norm < abs_error_tol:
+            del derivatives[key]
+            continue
+
+        above_abs, above_rel = _errors_above_tol(derivative_info, abs_error_tol, rel_error_tol)
+
+        if show_only_incorrect and not (above_abs or above_rel or inconsistent):
+            continue
+
+        yield key, fd_norm, fd_opts, directional, above_abs, above_rel, inconsistent
+
+
+def _format_error(error, tol):
+    """
+    Format the error, flagging if necessary.
+
+    Parameters
+    ----------
+    error : float
+        The absolute or relative error.
+    tol : float
+        Tolerance above which errors are flagged
+
+    Returns
+    -------
+    str
+        Formatted and possibly flagged error.
+    """
+    if np.isnan(error) or error < tol:
+        return f'{error:.6e}'
+    return f'{error:.6e} *'
